pyproject.toml
[project]
name = "recsub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg==0.30.0",
    "python-dotenv==1.1.0",
    "reflex>=0.8.0",
    "reflex-clerk-api>=1.0.1",
]
